fix: rewrite querySets to speed up SQL queries
The problem was because merges "join" with externalEID tables that is not using the foreignkeys ( because it I think works in only one direction e.g reactionEID table -> reaction table and not reaction -> reactionEID). A many to Many field should probably be used. The fix forces Django to build queries without join clauses. queries are split into subqueries by Postgres planner but is it still 300% faster.

+ remove .count() that were running some queries twice